#1679 add multi-layer support
