+++ GdaShape.h
@@ -202,7 +202,10 @@ protected:
 
 
 class GdaPoint: public GdaShape {
+
 public:
+    int radius;
+    
 	GdaPoint(); // creates a null shape
 	GdaPoint(const GdaPoint& s); 
 	GdaPoint(wxRealPoint point_o_s);
@@ -621,61 +624,5 @@ public:
 //	{ return lhs->z < rhs->z; }
 //};
 
-class BackgroundMapLayer
-{
-    Shapefile::ShapeType shape_type;
-    
-    wxColour pen_color;
-    wxColour brush_color;
-    int point_radius;
-    int opacity;
-    int pen_size;
-    bool show_boundary;
-    bool is_hide;
-    
-public:
-    vector<GdaShape*> shapes;
-    
-    BackgroundMapLayer(Shapefile::ShapeType shape_type,
-                       vector<GdaShape*> geoms);
-    ~BackgroundMapLayer();
-    
-    void SetHide(bool flag);
-    bool IsHide();
-    void SetPenColour(wxColour& color);
-    void SetBrushColour(wxColour& color);
-    void SetPointRadius(int radius);
-    void SetOpacity(int opacity);
-    void SetPenSize(int size);
-    void ShowBoundary(bool show);
-    wxColour GetBrushColour();
-    wxColour GetPenColour();
-    int GetPenSize();
-    int GetPointRadius();
-    int GetOpacity();
-    bool IsShowBoundary();
-    
-    vector<GdaShape*>& GetShapes();
-    Shapefile::ShapeType GetShapeType();
-    
-    void drawLegend(wxDC& dc, int x, int y, int w, int h);
-};
-
-class GdaShapeLayer : public GdaShape  {
-    wxString name;
-    BackgroundMapLayer* ml;
-    
-public:
-    GdaShapeLayer(wxString name, BackgroundMapLayer* ml);
-    ~GdaShapeLayer();
-
-    virtual GdaShape* clone();
-    virtual void Offset(double dx, double dy);
-    virtual void Offset(int dx, int dy);
-    virtual void applyScaleTrans(const GdaScaleTrans& A);
-    virtual void projectToBasemap(GDA::Basemap* basemap, double scale_factor = 1.0);
-    virtual void paintSelf(wxDC& dc);
-    virtual void paintSelf(wxGraphicsContext* gc);
-};
 
 #endif
